requirements.txt
torch
torchvision
xformers==0.0.32.post2  # torch==2.8+cu128
imageio
Pillow
imageio-ffmpeg
omegaconf
einops
transformers
torchmetrics
datasets
rich
wandb
opencv-python
opencv-python-headless
ipdb
tqdm
easydict
lpips
jaxtyping
scikit-image
numpy
